--- scripts/test-helper/src/path.ts
@@ -1,13 +1,0 @@
-import path from 'node:path';
-import upath from 'upath';
-
-export const isPathString = (test: string): boolean =>
-  path.posix.basename(test) !== test || path.win32.basename(test) !== test;
-
-export const isRelativePath = (test: string): boolean =>
-  /^\.\.?($|[\\/])/.test(test);
-
-export const normalizeToPosixPath = (p: string | undefined) =>
-  upath
-    .normalizeSafe(path.normalize(p || ''))
-    .replace(/^([a-zA-Z]+):/, (_, m: string) => `/${m.toLowerCase()}`);
